+++ src/components/ProtectedRoute.js
@@ -1,17 +1,18 @@
 import React from "react"
 import { Route, Redirect } from "react-router-dom"
 
-import { useAuth } from "../contexts/AuthContext"
+import { useAuth } from "../contexts/FirebaseAuth"
 
 export default function ProtectedRoute({ component: Component, ...rest }) {
-  const { currentUser } = useAuth()
+    const { currentUser } = useAuth()
 
-  return (
-    <Route
-      {...rest}
-      render={props => {
-        return currentUser ? <Component {...props} /> : <Redirect to="/login" />
-      }}
-    ></Route>
-  )
+    return (
+        <Route
+            {...rest}
+            render={props => {
+                return currentUser ? <Component {...props} /> : <Redirect to="/login" />
+            }}
+        >
+        </Route>
+    )
 }
