+++ src/firebase.js
@@ -15,6 +15,4 @@ const firebaseConfig = {
 
 const app = firebase.initializeApp(firebaseConfig)
 
-export const auth = app.auth()
-
 export default app
